refactor(components): extract layout primitives (Phase 6 #3)
Splits Card, Container, List, and Table out of builtins.tsx into
src/components/primitives/layout.tsx. builtins.tsx re-exports them
and drops the now-unused A2UIComponent type import.

No logic changes. Refs #43.
